requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
praw==7.7.1
textblob==0.17.1
vaderSentiment==3.3.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.24.3
wordcloud==1.9.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
gunicorn==21.2.0 